+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][unreleased]
 
+## [1.1.0] - 2020-04-04
+
+### Changed
+
+ - Modified output to match the GFM spec
+
+### Deprecated
+
+**This extension has been deprecated**.  All of its functionality now exists in league/commonmark 1.3+ under the `League\CommonMark\Extension\TaskList` namespace.
+
 ## [1.0.1] - 2019-07-10
 
 ### Changed
@@ -28,7 +38,8 @@ No code changes have been introduced since 1.0.0-beta1.
 
 Initial release!
 
-[unreleased]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v1.0.0-beta1...v1.0.0
 [1.0.0-beta1]: https://github.com/thephpleague/commonmark-ext-task-list/compare/v0.1.0...v1.0.0-beta1